+++ templates/content/blank/NewApp.WPF/App.fs
@@ -1,0 +1,21 @@
+﻿namespace NewApp.WPF
+
+open System
+
+open Xamarin.Forms
+open Xamarin.Forms.Platform.WPF
+
+type MainWindow() = 
+    inherit FormsApplicationPage()
+
+module Main = 
+    [<EntryPoint>]
+    [<STAThread>]
+    let main(_args) =
+
+        let app = new System.Windows.Application()
+        Forms.Init()
+        let window = MainWindow() 
+        window.LoadApplication(new NewApp.App())
+
+        app.Run(window)